@@ -61,8 +61,8 @@ data_files = [f[len(PACKAGENAME)+1:] for f in data_files]
 setup(name=PACKAGENAME,
       version=VERSION,
       description=DESCRIPTION,
-      python_requires='>=3.4',
-      install_requires=['numpy>=1.7.0',
+      python_requires='>=3.5',
+      install_requires=['numpy>=1.13.3',
                         'scipy>=0.9.0',
                         'extinction>=0.2.2',
                         'astropy>=1.0.0'],
