@@ -8,7 +8,7 @@ import os
 from astropy.config import ConfigItem, ConfigNamespace
 from astropy.config.configuration import update_default_config
 
-__version__ = "2.0.0"
+__version__ = "2.3.0"
 
 
 def test(package=None, test_path=None, args=None, plugins=None,
@@ -138,6 +138,7 @@ from .simulation import *
 from .plotting import *
 from .photdata import *
 from .registry import *
+from .specmodel import *
 
 # Register all the built-ins.
 from .builtins import *
