@@ -1,17 +1,16 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 
 import copy
-import time
 import math
-from collections import OrderedDict
+import time
 import warnings
+from collections import OrderedDict
 
 import numpy as np
-from scipy.interpolate import InterpolatedUnivariateSpline as Spline1d
 
-from .photdata import photometric_data, select_data
-from .utils import Result, Interp1D, ppf
-from .bandpasses import get_bandpass
+from .bandpasses import Bandpass
+from .photdata import photometric_data
+from .utils import Interp1D, Result, ppf
 
 __all__ = ['fit_lc', 'nest_lc', 'mcmc_lc', 'flatten_result', 'chisq']
 
@@ -20,7 +19,7 @@ class DataQualityError(Exception):
     pass
 
 
-def generate_chisq(data, model, signature='iminuit', modelcov=False):
+def generate_chisq(data, model, spectra, signature='iminuit', modelcov=False):
     """Define and return a chisq function for use in optimization.
 
     This function pre-computes and saves the inverse covariance matrix,
@@ -28,23 +27,55 @@ def generate_chisq(data, model, signature='iminuit', modelcov=False):
     is fixed at the time the chisq function is generated."""
 
     # precompute inverse covariance matrix
-    cov = np.diag(data.fluxerr**2) if data.fluxcov is None else data.fluxcov
-    if modelcov:
-        _, mcov = model.bandfluxcov(data.band, data.time,
-                                    zp=data.zp, zpsys=data.zpsys)
-        cov = cov + mcov
-    invcov = np.linalg.pinv(cov)
+    if data is not None:
+        cov = (np.diag(data.fluxerr**2) if data.fluxcov is None else
+               data.fluxcov)
+        if modelcov:
+            _, mcov = model.bandfluxcov(data.band, data.time,
+                                        zp=data.zp, zpsys=data.zpsys)
+            cov = cov + mcov
+        invcov = np.linalg.pinv(cov)
+
+    # If we have spectra, build the covariance matrix for them individually.
+    if spectra is not None:
+        if modelcov:
+            raise ValueError('modelcov not supported for spectra')
+
+        spectra = np.atleast_1d(spectra)
+        spectra_invcovs = []
+        for spectrum in spectra:
+            spectra_invcovs.append(np.linalg.pinv(spectrum.fluxcov))
 
     # iminuit expects each parameter to be a separate argument (including fixed
     # parameters)
     if signature == 'iminuit':
         def chisq(*parameters):
             model.parameters = parameters
-            model_flux = model.bandflux(data.band, data.time,
-                                        zp=data.zp, zpsys=data.zpsys)
-            diff = data.flux - model_flux
-            return np.dot(np.dot(diff, invcov), diff)
 
+            full_chisq = 0.
+
+            if data is not None:
+                model_flux = model.bandflux(data.band, data.time,
+                                            zp=data.zp, zpsys=data.zpsys)
+                diff = data.flux - model_flux
+                phot_chisq = np.dot(np.dot(diff, invcov), diff)
+                full_chisq += phot_chisq
+
+            if spectra is not None:
+                for spectrum, spec_invcov in zip(spectra, spectra_invcovs):
+                    sample_wave, sampling_matrix = \
+                        spectrum.get_sampling_matrix()
+                    sample_flux = model.flux(spectrum.time, sample_wave)
+                    spec_model_flux = (
+                        sampling_matrix.dot(sample_flux) /
+                        sampling_matrix.dot(np.ones_like(sample_flux))
+                    )
+                    spec_diff = spectrum.flux - spec_model_flux
+                    spec_chisq = spec_invcov.dot(spec_diff).dot(spec_diff)
+
+                    full_chisq += spec_chisq
+
+            return full_chisq
     else:
         raise ValueError("unknown signature: {!r}".format(signature))
 
@@ -130,13 +161,13 @@ def flatten_result(res):
     for n1 in res.param_names:
         for n2 in res.param_names:
             key = n1 + '_' + n2 + '_cov'
-            if n1 not in res.cov_names or n2 not in res.cov_names:
+            if n1 not in res.vparam_names or n2 not in res.vparam_names:
                 flat[key] = 0.
             elif res.covariance is None:
                 flat[key] = float('nan')
             else:
-                i = res.cov_names.index(n1)
-                j = res.cov_names.index(n2)
+                i = res.vparam_names.index(n1)
+                j = res.vparam_names.index(n2)
                 flat[key] = res.covariance[i, j]
 
     return flat
@@ -175,19 +206,30 @@ def cut_bands(data, model, z_bounds=None, warn=True):
     return data, mask
 
 
-def t0_bounds(data, model):
+def t0_bounds(data, model, spectra=None):
     """Determine bounds on t0 parameter of the model.
 
     The lower bound is such that the latest model time is equal to the
     earliest data time. The upper bound is such that the earliest
     model time is equal to the latest data time."""
 
-    return (model.get('t0') + np.min(data.time) - model.maxtime(),
-            model.get('t0') + np.max(data.time) - model.mintime())
+    times = []
+
+    if data is not None:
+        times.append(data.time)
+
+    if spectra is not None:
+        for spec in np.atleast_1d(spectra):
+            times.append(spec.time)
+
+    times = np.hstack(times)
+
+    return (model.get('t0') + np.min(times) - model.maxtime(),
+            model.get('t0') + np.max(times) - model.mintime())
 
 
-def guess_t0_and_amplitude(data, model, minsnr):
-    """Guess t0 and amplitude of the model based on the data."""
+def _guess_t0_and_amplitude_photometry(data, model, minsnr):
+    """Guess t0 and amplitude of the model from photometry."""
 
     # get data above the signal-to-noise ratio cut
     significant_data = data[(data.flux / data.fluxerr) > minsnr]
@@ -238,13 +280,86 @@ def guess_t0_and_amplitude(data, model, minsnr):
     return t0, amplitude
 
 
-def _print_iminuit_params(names, kwargs):
+def _guess_t0_and_amplitude_spectra(spectra, model, minsnr):
+    """Guess t0 and amplitude of the model from spectra.
+
+    The spectra don't necessarily have the same binning which makes this
+    challenging. To handle this, we synthesize photometry in a range of
+    different synthetic filters. We then call
+    `_guess_t0_and_amplitude_photometry` on this photometry.
+    """
+
+    # Build a set of bands to use for synthetic photometry.
+    target_band_width = 500     # Angstroms
+    minwave = model.minwave()
+    maxwave = model.maxwave()
+    band_count = int(math.ceil((maxwave - minwave) / target_band_width))
+    band_edges = np.linspace(minwave, maxwave, band_count+1)
+    band_starts = band_edges[:-1]
+    band_ends = band_edges[1:]
+
+    bandpasses = np.array([Bandpass([start, end], [1., 1.]) for start, end
+                           in zip(band_starts, band_ends)])
+
+    all_bands = []
+    all_fluxes = []
+    all_fluxerrs = []
+    all_times = []
+
+    for spectrum in np.atleast_1d(spectra):
+        # Find the bandpasses that overlap this spectrum.
+        band_mask = ((spectrum.bin_edges[0] <= band_starts)
+                     & (spectrum.bin_edges[-1] >= band_ends))
+        spec_bands = bandpasses[band_mask]
+
+        spec_flux, spec_fluxcov = spectrum.bandfluxcov(spec_bands, zp=25.,
+                                                       zpsys='ab')
+        spec_fluxerr = np.sqrt(np.diag(spec_fluxcov))
+
+        all_bands.extend(spec_bands)
+        all_fluxes.extend(spec_flux)
+        all_fluxerrs.extend(spec_fluxerr)
+        all_times.extend(np.ones_like(spec_flux) * spectrum.time)
+
+    photometry = photometric_data({
+        'band': all_bands,
+        'flux': all_fluxes,
+        'fluxerr': all_fluxerrs,
+        'time': all_times,
+        'zp': [25.] * len(all_fluxes),
+        'zpsys': ['ab'] * len(all_fluxes),
+    })
+
+    # Sort the photometry by time
+    photometry = photometry[np.argsort(photometry.time)]
+
+    return _guess_t0_and_amplitude_photometry(photometry, model, minsnr)
+
+
+def guess_t0_and_amplitude(data, model, minsnr, spectra=None):
+    """Guess t0 and amplitude of the model based on the data.
+
+    If we have photometry, we use it for the guessing. If not, we use all
+    available spectra instead.
+    """
+    if data is not None:
+        return _guess_t0_and_amplitude_photometry(data, model, minsnr)
+    elif spectra is not None:
+        return _guess_t0_and_amplitude_spectra(spectra, model, minsnr)
+    else:
+        raise ValueError('need either photometry or spectra to guess t0 and '
+                         'amplitude.')
+
+
+def _print_iminuit_params(names, fixed_parameters, start_values,
+                          start_errors, bounds):
     """Verbose printing of parameters to pass to Minuit"""
     for name in names:
-        print(name, kwargs[name], 'step=', kwargs['error_' + name],
-              end=" ")
-        if 'limit_' + name in kwargs:
-            print('bounds=', kwargs['limit_' + name], end=" ")
+        if name in fixed_parameters:
+            continue
+        print(name, start_values[name], 'step=', start_errors[name], end=" ")
+        if name in bounds:
+            print("bounds=", bounds[name], end=" ")
         print()
 
 
@@ -271,8 +386,81 @@ def _phase_and_wave_mask(data, t0, z, phase_range, wave_range):
     return None
 
 
-def fit_lc(data, model, vparam_names, bounds=None, method='minuit',
-           guess_amplitude=True, guess_t0=True, guess_z=True,
+def _run_iminuit(chisq, parameter_names, start_values, start_errors, bounds,
+                 fixed_parameters, maxcall, verbose=False):
+    try:
+        import iminuit
+    except ImportError:
+        raise ValueError("Minimization method 'minuit' requires the iminuit "
+                         "package")
+
+    # The iminuit API changed significantly in version 2. Handle both the new
+    # and old APIs.
+    from distutils.version import LooseVersion
+    iminuit_version = LooseVersion(iminuit.__version__)
+
+    if verbose:
+        print("Initial parameters:")
+        _print_iminuit_params(parameter_names, fixed_parameters, start_values,
+                              start_errors, bounds)
+        print()
+
+    if iminuit_version >= LooseVersion("2.0.0"):
+        # New API
+        m = iminuit.Minuit(chisq, name=parameter_names, **start_values)
+
+        m.errordef = iminuit.Minuit.LEAST_SQUARES
+        m.print_level = 1 if verbose >= 2 else 0
+        m.throw_nan = True
+
+        for key, value in start_errors.items():
+            m.errors[key] = value
+
+        for key, value in bounds.items():
+            m.limits[key] = value
+
+        for key in fixed_parameters:
+            m.fixed[key] = True
+            m.errors[key] = 0.
+
+        m.migrad(ncall=maxcall)
+        fmin = m.fmin
+    else:
+        # Old API
+
+        # Set up keyword arguments to pass to Minuit initializer.
+        kwargs = {}
+
+        # forced_parameters/name keyword also changed in v1.4.3, handle that.
+        if iminuit_version >= LooseVersion("1.4.3"):
+            kwargs['name'] = parameter_names
+        else:
+            kwargs['forced_parameters'] = parameter_names
+
+        for key, value in start_values.items():
+            kwargs[key] = value
+
+        for key, value in start_errors.items():
+            kwargs['error_' + key] = value
+
+        for key, value in bounds.items():
+            kwargs['limit_' + key] = value
+
+        for key in fixed_parameters:
+            kwargs['fix_' + key] = True
+            kwargs['error_' + key] = 0.
+
+        m = iminuit.Minuit(chisq, errordef=1.,
+                           print_level=(1 if verbose >= 2 else 0),
+                           throw_nan=True, **kwargs)
+
+        fmin, params = m.migrad(ncall=maxcall)
+
+    return m, fmin
+
+
+def fit_lc(data=None, model=None, vparam_names=[], bounds=None, spectra=None,
+           method='minuit', guess_amplitude=True, guess_t0=True, guess_z=True,
            minsnr=5.0, modelcov=False, verbose=False, maxcall=10000,
            phase_range=None, wave_range=None, warn=True):
     """Fit model parameters to data by minimizing chi^2.
@@ -298,6 +486,8 @@ def fit_lc(data, model, vparam_names, bounds=None, method='minuit',
         phase of the model lines up with the earliest data point and
         the maximum bound is such that the earliest phase of the model
         lines up with the latest data point.
+    spectra : `~sncosmo.Spectrum` or list of `~sncosmo.Spectrum` objects
+        A list of spectra to include in the fit.
     guess_amplitude : bool, optional
         Whether or not to guess the amplitude from the data. If false, the
         current model amplitude is taken as the initial value. Only has an
@@ -411,18 +601,30 @@ def fit_lc(data, model, vparam_names, bounds=None, method='minuit',
 
     """
 
-    # Standardize data
-    data = photometric_data(data)
+    if data is not None:
+        # Standardize data
+        data = photometric_data(data)
 
-    # sort by time (some sources require this)
-    # We keep track of indicies that sort the array because we want to be
-    # able to report the indexes of the original data that were used in the
-    # fit.
-    if not np.all(np.ediff1d(data.time) >= 0.0):
-        sortidx = np.argsort(data.time)
-        data = data[sortidx]
+        # sort by time (some sources require this)
+        # We keep track of indicies that sort the array because we want to be
+        # able to report the indexes of the original data that were used in the
+        # fit.
+        if not np.all(np.ediff1d(data.time) >= 0.0):
+            sortidx = np.argsort(data.time)
+            data = data[sortidx]
+        else:
+            sortidx = None
     else:
         sortidx = None
+
+    if spectra is not None:
+        # Make sure that we have times for all of the spectra
+        for spectrum in np.atleast_1d(spectra):
+            if spectrum.time is None:
+                raise TypeError("'time' must be set for each spectrum")
+
+    if model is None:
+        raise TypeError("missing required argument 'model'")
 
     # Make a copy of the model so we can modify it with impunity.
     model = copy.copy(model)
@@ -439,8 +641,7 @@ def fit_lc(data, model, vparam_names, bounds=None, method='minuit',
     vparam_names = [s for s in model.param_names if s in vparam_names]
 
     # initialize bounds
-    if bounds is None:
-        bounds = {}
+    bounds = copy.deepcopy(bounds) if bounds else {}
 
     # Check that 'z' is bounded (if it is going to be fit).
     if 'z' in vparam_names:
@@ -451,18 +652,21 @@ def fit_lc(data, model, vparam_names, bounds=None, method='minuit',
         if model.get('z') < bounds['z'][0] or model.get('z') > bounds['z'][1]:
             raise ValueError('z out of range.')
 
-    # Cut bands that are not allowed by the wavelength range of the model.
-    fitdata, support_mask = cut_bands(data, model,
-                                      z_bounds=bounds.get('z', None),
-                                      warn=warn)
-    data_mask = support_mask  # Initially this is the complete mask on data.
-
-    # Unique set of bands in data
-    bands = set(fitdata.band.tolist())
+    if data is not None:
+        # Cut bands that are not allowed by the wavelength range of the model.
+        fitdata, support_mask = cut_bands(data, model,
+                                          z_bounds=bounds.get('z', None),
+                                          warn=warn)
+        # Initially this is the complete mask on data.
+        data_mask = support_mask
+    else:
+        fitdata = None
+        support_mask = None
+        data_mask = None
 
     # Find t0 bounds to use, if not explicitly given
     if 't0' in vparam_names and 't0' not in bounds:
-        bounds['t0'] = t0_bounds(fitdata, model)
+        bounds['t0'] = t0_bounds(fitdata, model, spectra)
 
     # Note that in the parameter guessing below, we assume that the source
     # amplitude is the 3rd parameter of the Model (1st parameter of the Source)
@@ -476,39 +680,31 @@ def fit_lc(data, model, vparam_names, bounds=None, method='minuit',
     # Make guesses for t0 and amplitude.
     # (For now, we assume it is the 3rd parameter of the model.)
     if (guess_amplitude or guess_t0):
-        t0, amplitude = guess_t0_and_amplitude(fitdata, model, minsnr)
+        t0, amplitude = guess_t0_and_amplitude(fitdata, model, minsnr, spectra)
         if guess_amplitude:
             model.parameters[2] = amplitude
         if guess_t0:
             model.set(t0=t0)
 
-    # count degrees of freedom
-    ndof = len(fitdata) - len(vparam_names)
-
     if method == 'minuit':
-        try:
-            import iminuit
-        except ImportError:
-            raise ValueError("Minimization method 'minuit' requires the "
-                             "iminuit package")
+        fixed_parameters = []
+        start_values = {}
+        start_errors = {}
 
-        # Set up keyword arguments to pass to Minuit initializer.
-        kwargs = {}
+        # Figure out parameters to pass to iminuit.
         for name in model.param_names:
-            kwargs[name] = model.get(name)  # Starting point.
+            start_values[name] = model.get(name)
 
             # Fix parameters not being varied in the fit.
             if name not in vparam_names:
-                kwargs['fix_' + name] = True
-                kwargs['error_' + name] = 0.
+                fixed_parameters.append(name)
                 continue
 
-            # Bounds
+            # Check that the bounds are valid.
             if name in bounds:
                 if None in bounds[name]:
                     raise ValueError('one-sided bounds not allowed for '
                                      'minuit minimizer')
-                kwargs['limit_' + name] = bounds[name]
 
             # Initial step size
             if name in bounds:
@@ -517,26 +713,28 @@ def fit_lc(data, model, vparam_names, bounds=None, method='minuit',
                 step = 0.1 * model.get(name)
             else:
                 step = 1.
-            kwargs['error_' + name] = step
+            start_errors[name] = step
 
-        if verbose:
-            print("Initial parameters:")
-            _print_iminuit_params(vparam_names, kwargs)
-            print()
+        # count degrees of freedom
+        ndof = 0
+        if data is not None:
+            ndof += len(fitdata)
+        elif spectra is not None:
+            for spectrum in np.atleast_1d(spectra):
+                ndof += len(spectrum)
+        ndof -= len(vparam_names)
 
         # run once with no model covariance, regardless of whether
         # modelcov=True
-        fitchisq = generate_chisq(fitdata, model, signature='iminuit',
+        fitchisq = generate_chisq(fitdata, model, spectra, signature='iminuit',
                                   modelcov=False)
-        ndof = len(fitdata) - len(vparam_names)
 
-        m = iminuit.Minuit(fitchisq, errordef=1.,
-                           forced_parameters=model.param_names,
-                           print_level=(1 if verbose >= 2 else 0),
-                           throw_nan=True, **kwargs)
-        d, l = m.migrad(ncall=maxcall)
+        m, fmin = _run_iminuit(fitchisq, model.param_names, start_values,
+                               start_errors, bounds, fixed_parameters, maxcall,
+                               verbose)
+
         if verbose:
-            print("{} function calls; {} dof.".format(d.nfcn, ndof))
+            print("{} function calls; {} dof.".format(fmin.nfcn, ndof))
 
         # numpy array of best-fit values (including fixed parameters).
         parameters = np.array([m.values[name] for name in model.param_names])
@@ -545,6 +743,9 @@ def fit_lc(data, model, vparam_names, bounds=None, method='minuit',
         # Iterative Fitting
 
         if phase_range or wave_range:
+            if spectra is not None:
+                raise ValueError('phase_range and wave_range are not '
+                                 'supported for spectra')
             range_mask = _phase_and_wave_mask(data, model.get('t0'),
                                               model.get('z'),
                                               phase_range, wave_range)
@@ -555,35 +756,36 @@ def fit_lc(data, model, vparam_names, bounds=None, method='minuit',
         # masked points.
         refit = (modelcov or ((phase_range or wave_range) and
                               np.any(data_mask != support_mask)))
+
         nfit = 1
         while refit:
             # set new starting point to last point
             for name in vparam_names:
-                kwargs[name] = model.get(name)
-
-            if verbose:
-                print("Initial parameters:")
-                _print_iminuit_params(vparam_names, kwargs)
-                print()
+                start_values[name] = model.get(name)
 
             # re-crop data based on ranges, if necessary
             if (phase_range or wave_range):
                 fitdata = data[data_mask]
 
-            ndof = len(fitdata) - len(vparam_names)
+            # count degrees of freedom
+            ndof = 0
+            if data is not None:
+                ndof += len(fitdata)
+            elif spectra is not None:
+                for spectrum in np.atleast_1d(spectra):
+                    ndof += len(spectrum)
+            ndof -= len(vparam_names)
 
             # generate chisq function based on new starting point
-            fitchisq = generate_chisq(fitdata, model, signature='iminuit',
-                                      modelcov=modelcov)
+            fitchisq = generate_chisq(fitdata, model, spectra,
+                                      signature='iminuit', modelcov=modelcov)
 
-            m = iminuit.Minuit(fitchisq, errordef=1.,
-                               forced_parameters=model.param_names,
-                               print_level=(1 if verbose >= 2 else 0),
-                               throw_nan=True, **kwargs)
-            d, l = m.migrad(ncall=maxcall)
+            m, fmin = _run_iminuit(fitchisq, model.param_names, start_values,
+                                   start_errors, bounds, fixed_parameters,
+                                   maxcall, verbose)
 
             if verbose:
-                print("{} function calls; {} dof.".format(d.nfcn, ndof))
+                print("{} function calls; {} dof.".format(fmin.nfcn, ndof))
 
             parameters = np.array([m.values[name]
                                    for name in model.param_names])
@@ -593,7 +795,7 @@ def fit_lc(data, model, vparam_names, bounds=None, method='minuit',
             refit = False
             # only consider refitting if we got a valid answer and we're under
             # the maximum number of iterations:
-            if d.is_valid and nfit < 22:
+            if fmin.is_valid and nfit < 22:
                 # recalculate data mask based on new t0, z
                 if phase_range or wave_range:
                     old_data_mask = data_mask
@@ -609,25 +811,25 @@ def fit_lc(data, model, vparam_names, bounds=None, method='minuit',
                 # statistical error bar
                 if modelcov:
                     for name in vparam_names:
-                        frac_change = (abs(m.values[name] - kwargs[name]) /
-                                       m.errors[name])
+                        frac_change = (abs(m.values[name] - start_values[name])
+                                       / m.errors[name])
                         refit = refit or frac_change > 0.1
 
         # Build a message.
         message = []
-        if d.has_reached_call_limit:
+        if fmin.has_reached_call_limit:
             message.append('Reached call limit.')
-        if d.hesse_failed:
+        if fmin.hesse_failed:
             message.append('Hesse Failed.')
-        if not d.has_covariance:
+        if not fmin.has_covariance:
             message.append('No covariance.')
-        elif not d.has_accurate_covar:  # iminuit docs wrong
+        elif not fmin.has_accurate_covar:  # iminuit docs wrong
             message.append('Covariance may not be accurate.')
-        if not d.has_posdef_covar:  # iminuit docs wrong
+        if not fmin.has_posdef_covar:  # iminuit docs wrong
             message.append('Covariance not positive definite.')
-        if d.has_made_posdef_covar:
+        if fmin.has_made_posdef_covar:
             message.append('Covariance forced positive definite.')
-        if not d.has_valid_parameters:
+        if not fmin.has_valid_parameters:
             message.append('Parameter(s) value and/or error invalid.')
         if len(message) == 0:
             message.append('Minimization exited successfully.')
@@ -654,10 +856,10 @@ def fit_lc(data, model, vparam_names, bounds=None, method='minuit',
             data_mask = data_mask[unsort_idx]
 
         # Compile results
-        res = Result(success=d.is_valid,
+        res = Result(success=fmin.is_valid,
                      message=' '.join(message),
-                     ncall=d.nfcn,
-                     chisq=d.fval,
+                     ncall=fmin.nfcn,
+                     chisq=fmin.fval,
                      ndof=ndof,
                      param_names=model.param_names,
                      parameters=parameters,
@@ -1067,8 +1269,7 @@ def mcmc_lc(data, model, vparam_names, bounds=None, priors=None,
     # Make a copy of the model so we can modify it with impunity.
     model = copy.copy(model)
 
-    if bounds is None:
-        bounds = {}
+    bounds = copy.deepcopy(bounds) if bounds else {}
     if priors is None:
         priors = {}
 
