@@ -4,13 +4,11 @@ import abc
 import math
 
 import numpy as np
-import astropy.units as u
-import astropy.constants as const
 
 from ._registry import Registry
 from .bandpasses import get_bandpass
-from .utils import integration_grid
 from .constants import H_ERG_S, SPECTRUM_BANDFLUX_SPACING
+from .utils import integration_grid
 
 __all__ = ['get_magsystem', 'MagSystem', 'SpectralMagSystem',
            'ABMagSystem', 'CompositeMagSystem']
@@ -169,7 +167,7 @@ class SpectralMagSystem(MagSystem):
 
     Parameters
     ----------
-    refspectrum : `sncosmo.Spectrum`
+    refspectrum : `sncosmo.SpectrumModel`
         The spectrum of the fundamental spectrophotometric standard.
     """
 
