@@ -3,10 +3,9 @@
 import math
 
 import numpy as np
-from numpy.testing import assert_allclose, assert_almost_equal
-from astropy import units as u
-from astropy.utils.data import get_pkg_data_filename
 import pytest
+from astropy import units as u
+from numpy.testing import assert_allclose, assert_almost_equal
 
 import sncosmo
 
@@ -24,11 +23,11 @@ def test_spectralmagsystem():
 
     # construct a spectrum with same flux as AB: 3631 x 10^{-23} erg/s/cm^2/Hz
     # Use a fine grid to reduce linear interpolation errors when integrating
-    # in Spectrum.bandflux().
+    # in SpectrumModel.bandflux().
     wave = np.linspace(1000., 20000., 100000)  # fine grid
     flux = 3631.e-23 * np.ones_like(wave)
     unit = u.erg / u.s / u.cm**2 / u.Hz
-    s = sncosmo.Spectrum(wave, flux, unit=unit)
+    s = sncosmo.SpectrumModel(wave, flux, unit=unit)
     magsys1 = sncosmo.SpectralMagSystem(s)
 
     magsys2 = sncosmo.ABMagSystem()
